move / to /home. move /splash to /
